Truth-Lens-AI-main/Github-Upload-Folder/backend/requirements.txt
# TruthLens AI — Python Dependencies
# Install with: pip install -r requirements.txt

# ── Web Framework ──────────────────────────────────────────
fastapi>=0.111.0
uvicorn[standard]>=0.29.0
python-multipart>=0.0.9

# ── AI / ML ────────────────────────────────────────────────
scikit-learn>=1.5.0
numpy>=1.26.4
scipy>=1.13.0
joblib>=1.4.2

# ── Data
pandas>=2.2.0

# ── Image processing ───────────────────────────────────────
Pillow>=10.3.0
opencv-python-headless>=4.9.0

# ── NLP ────────────────────────────────────────────────────
nltk>=3.8.1
textblob>=0.18.0

# ── Utilities ──────────────────────────────────────────────
tldextract>=5.1.2
requests>=2.31.0
aiofiles>=23.2.1
python-dotenv>=1.0.1

# ── AI Providers ───────────────────────────────────────────
google-generativeai>=0.5.4

# ── Optional: only needed if deepfake_model.pth exists ────
# torch>=2.3.0
# torchvision>=0.18.0
